Fixed vertex lighting in GLSL for AMD/ATI cards
Ended up using SSBO, because neither TBO nor UBO worked with any AMD gpus we tested.
SSBO is only supported since 2012, but that shouldn't be a problem now. I had to get rid of the GLSL 1.4 compatibility profile and update the shader entirely for GLSL 4.3
Also fixed fog rendering in GLSL. Todo: make common code for calculating fog parameters (start, end, density) to avoid code duplication. Also would be ideal to put all fog parameters to a uniform block.

+++ trunk/gl_fog.c
@@ -326,7 +326,7 @@ void Fog_SetupFrame (void)
 	glFogfv(GL_FOG_COLOR, Fog_GetColor());
 	if (Fog_IsWaterFog())
 	{
-		if (((int)gl_waterfog.value) == 2)
+		if (gl_waterfog.value == 2)
 		{
 			glFogi(GL_FOG_MODE, GL_EXP);
 			glFogf(GL_FOG_DENSITY, 0.0002 + (0.0009 - 0.0002) * bound(0, gl_waterfog_density.value, 1));
